dodanie pol do firmy

+++ src/main/java/com/tackpad/dao/MessageDao.java
@@ -5,7 +5,6 @@ package com.tackpad.dao;
 import com.tackpad.models.Message;
 import com.tackpad.models.enums.MessageType;
 import com.tackpad.requests.enums.ListingSortType;
-import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.text.ParseException;
 import java.util.List;
@@ -15,5 +14,5 @@ public interface MessageDao extends BaseDao<Message> {
 						  List<Long> companyCategoryMainIdList, List<MessageType> messageTypeList, Double latitude,
 						  Double longitude, Integer range, String searchTerm, ListingSortType listingSortType) throws ParseException;
 
-    Integer findCount(Long companyId);
+    Long findCount(Long companyId);
 }
